pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heymax-analytics-stack"
version = "1.0.0"
description = "End-to-end analytics stack for HeyMax using Dagster, dbt, and BigQuery"
authors = [
    {name = "Edwin", email = "echiang95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "dagster>=1.5.13",
    "dagster-webserver>=1.5.13", 
    "dagster-postgres>=0.21.13",
    "dagster-gcp>=0.21.13",
    "dagster-dbt>=0.21.13",
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-storage>=2.10.0",
    "polars>=0.20.2",
    "dbt-bigquery>=1.6.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dagster_project*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.dagster]
module_name = "dagster_project"